fix: move close/menu buttons to header from nav
