+++ src/components/Header/Header.js
@@ -1,14 +1,28 @@
-import React from 'react'
+import { React, useState } from 'react'
 import Navigation from '../Navigation/Navigation'
 import { 
   Logo,
   Wrapper,
+  LogoMenuWrapper,
+  CloseButton,
+  MenuButton,
 } from './styledHeader'
 
 export default function Header({ loggedIn }) {
+
+  const [isOpen, setIsOpen] = useState(false);
+
   return (
     <Wrapper>
-      <Logo to='/'>NewsExplorer</Logo>
+      <LogoMenuWrapper>
+        <Logo to='/'>NewsExplorer</Logo>
+        {
+          isOpen ? 
+            <CloseButton onClick={() => setIsOpen(!isOpen)} /> 
+          : 
+            <MenuButton onClick={() => setIsOpen(!isOpen)} />
+        }
+      </LogoMenuWrapper>
       <Navigation loggedIn={loggedIn} />
     </Wrapper>
   )
